Add initial code into Task for notice
